# protocol/openapi.yaml
openapi: 3.1.0
info:
  title: Lenora Backend Protocol
  version: "1"
  license: { name: Apache-2.0, identifier: Apache-2.0 }
  description: >
    Contract between the Lenora editor and any Lenora backend. Clients ignore
    unknown fields and unknown kinds. Breaking changes move to /v2.
servers:
  - url: http://127.0.0.1:8787
security:
  - bearer: []
paths:
  /v1/health:
    get:
      operationId: getHealth
      security: [{}, { bearer: [] }]
      description: Without a valid token only status and protocolVersion are returned.
      parameters:
        - name: recheck
          in: query
          required: false
          description: "`addons` asks adapters to forget learned add-on unavailability before answering. Ignored without a valid token."
          schema: { enum: [addons] }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Health" } } } }
  /v1/capabilities:
    get:
      operationId: getCapabilities
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Capabilities" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /v1/uploads:
    post:
      operationId: createUpload
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UploadRequest" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/UploadTicket" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /v1/jobs:
    post:
      operationId: submitJob
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 1, maxLength: 255 }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/JobRequest" } } }
      responses:
        "202": { description: Accepted, content: { application/json: { schema: { $ref: "#/components/schemas/SubmittedJob" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /v1/jobs/{jobId}:
    parameters:
      - { name: jobId, in: path, required: true, schema: { type: string } }
    get:
      operationId: getJob
      responses:
        "200":
          description: OK
          headers:
            Retry-After: { description: Seconds until the next poll; present while not terminal., schema: { type: integer } }
          content: { application/json: { schema: { $ref: "#/components/schemas/JobState" } } }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      operationId: cancelJob
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/JobState" } } } }
        default: { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
  responses:
    Problem:
      description: RFC 9457 problem
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
  schemas:
    Kind:
      type: string
      description: Operation kind. Clients ignore kinds they do not recognise.
      examples: [image.removeBackground, image.generate, image.edit, image.upscale, video.generate, video.reframe, video.edit, video.lipSync, video.upscale, audio.speech, audio.music, audio.sfx, text.rewritePrompt]
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    ErrorCode:
      type: string
      enum: [unauthorized, invalid_request, unsupported_kind, unknown_model, input_too_large, rate_limited, quota_exceeded, provider_unavailable, provider_error, not_found, not_cancellable]
    Estimate:
      type: object
      required: [amount, unit]
      properties:
        amount: { type: number, minimum: 0 }
        unit: { type: string, examples: [cloudinary_credits, usd] }
    AdapterHealth:
      type: object
      required: [id, enabled]
      properties:
        id: { type: string }
        enabled: { type: boolean }
        reason: { type: [string, "null"] }
        details: { type: [object, "null"], description: "Adapter-specific state, for example add-on availability and budget use." }
    Health:
      type: object
      required: [status, protocolVersion]
      properties:
        status: { const: ok }
        protocolVersion: { const: "1" }
        backendVersion: { type: [string, "null"] }
        adapters: { type: [array, "null"], items: { $ref: "#/components/schemas/AdapterHealth" } }
    InputLimits:
      type: object
      required: [types, maxBytes]
      properties:
        types: { type: array, items: { type: string }, description: Accepted MIME types }
        maxBytes: { type: integer, minimum: 1 }
        maxPixels: { type: [integer, "null"], minimum: 1, description: Largest width × height accepted for image inputs. }
    ModelInfo:
      type: object
      required: [id, kind, displayName, inputs, cancellable]
      properties:
        id: { type: string, pattern: "^[a-z0-9_-]+/.+$", description: "<adapterId>/<model>" }
        kind: { $ref: "#/components/schemas/Kind" }
        displayName: { type: string }
        inputs: { $ref: "#/components/schemas/InputLimits" }
        cancellable: { type: boolean }
        estimate: { oneOf: [{ $ref: "#/components/schemas/Estimate" }, { type: "null" }] }
        ui: { type: [object, "null"], description: Optional editor hints (durations, aspects, settings) for generation kinds. }
        operations: { type: [array, "null"], items: { type: string }, description: "For image.edit: the ops this model performs." }
    AdapterVersion:
      type: object
      required: [id, version]
      properties:
        id: { type: string }
        version: { type: string }
    Capabilities:
      type: object
      required: [protocolVersion, adapters, models]
      properties:
        protocolVersion: { const: "1" }
        adapters: { type: array, items: { $ref: "#/components/schemas/AdapterVersion" } }
        models: { type: array, items: { $ref: "#/components/schemas/ModelInfo" } }
    UploadRequest:
      type: object
      required: [model, contentType, byteCount, filename]
      properties:
        model: { type: string }
        contentType: { type: string }
        byteCount: { type: integer, minimum: 1 }
        filename: { type: string, minLength: 1, maxLength: 255 }
    Ticket:
      type: object
      required: [method, url, headers, fields, expiresAt]
      properties:
        method: { enum: [POST, PUT] }
        url: { type: string, format: uri }
        headers: { type: object, additionalProperties: { type: string } }
        fields: { type: object, additionalProperties: { type: string }, description: Multipart form fields for POST; empty for PUT. }
        fileField: { type: [string, "null"], description: Multipart field name for the file; null for PUT. }
        expiresAt: { type: string, format: date-time }
    UploadTicket:
      type: object
      required: [assetRef, ticket]
      properties:
        assetRef: { type: string }
        ticket: { $ref: "#/components/schemas/Ticket" }
    InputRole:
      enum: [startFrame, endFrame, reference, null]
      description: "video.generate labels every input; image.generate inputs are references; other kinds leave it null."
    JobInput:
      oneOf:
        - { type: object, required: [assetRef], properties: { assetRef: { type: string }, role: { $ref: "#/components/schemas/InputRole" } }, additionalProperties: false }
        - { type: object, required: [url], properties: { url: { type: string, format: uri }, role: { $ref: "#/components/schemas/InputRole" } }, additionalProperties: false }
    JobRequest:
      type: object
      required: [kind, model]
      properties:
        kind: { $ref: "#/components/schemas/Kind" }
        model: { type: string }
        inputs: { type: array, maxItems: 8, default: [], items: { $ref: "#/components/schemas/JobInput" } }
        params: { type: object, default: {}, description: "Shape depends on kind; see the *Params schemas." }
    SubmittedJob:
      type: object
      required: [jobId, status]
      properties:
        jobId: { type: string }
        status: { $ref: "#/components/schemas/JobStatus" }
        estimate: { oneOf: [{ $ref: "#/components/schemas/Estimate" }, { type: "null" }] }
    JobResult:
      type: object
      required: [url, contentType, fileExtension]
      properties:
        url: { type: string, format: uri }
        contentType: { type: string }
        fileExtension: { type: string, pattern: "^[a-z0-9]+$" }
    JobError:
      type: object
      required: [code, message, retryable]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message: { type: string }
        retryable: { type: boolean }
    JobState:
      type: object
      required: [jobId, status]
      properties:
        jobId: { type: string }
        status: { $ref: "#/components/schemas/JobStatus" }
        results: { type: [array, "null"], items: { $ref: "#/components/schemas/JobResult" } }
        error: { oneOf: [{ $ref: "#/components/schemas/JobError" }, { type: "null" }] }
    Problem:
      type: object
      required: [type, title, status, code, retryable]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: [string, "null"] }
        code: { $ref: "#/components/schemas/ErrorCode" }
        retryable: { type: boolean }
    UrlPrompt:
      type: string
      pattern: "^[A-Za-z0-9 .'-]{1,100}$"
      description: Text placed inside a transformation URL.
    ImageAspect:
      enum: ["1:1", "16:9", "9:16", "4:3", "3:4"]
    RemoveBackgroundParams:
      type: object
      additionalProperties: false
    ImageUpscaleParams:
      type: object
      additionalProperties: false
    ImageGenerateParams:
      type: object
      additionalProperties: false
      required: [prompt]
      properties:
        prompt: { type: string, minLength: 1, maxLength: 1000 }
        aspectRatio: { $ref: "#/components/schemas/ImageAspect", default: "1:1" }
        count: { type: integer, minimum: 1, maximum: 4, default: 1 }
        seed: { type: [integer, "null"], minimum: 0, maximum: 2147483647 }
    ImageEditParams:
      oneOf:
        - { type: object, additionalProperties: false, required: [op, aspectRatio], properties: { op: { const: fill }, aspectRatio: { $ref: "#/components/schemas/ImageAspect" } } }
        - { type: object, additionalProperties: false, required: [op, from, to], properties: { op: { const: replace }, from: { $ref: "#/components/schemas/UrlPrompt" }, to: { $ref: "#/components/schemas/UrlPrompt" } } }
        - { type: object, additionalProperties: false, required: [op, prompt], properties: { op: { const: remove }, prompt: { $ref: "#/components/schemas/UrlPrompt" } } }
        - { type: object, additionalProperties: false, required: [op, prompt, color], properties: { op: { const: recolor }, prompt: { $ref: "#/components/schemas/UrlPrompt" }, color: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" } } }
        - { type: object, additionalProperties: false, required: [op], properties: { op: { const: backgroundReplace }, prompt: { oneOf: [{ $ref: "#/components/schemas/UrlPrompt" }, { type: "null" }] } } }
        - { type: object, additionalProperties: false, required: [op], properties: { op: { const: restore } } }
    VideoGenerateParams:
      type: object
      additionalProperties: false
      required: [prompt, duration]
      description: "1080p requires duration 8. Inputs: startFrame ≤1, endFrame ≤1 (only with startFrame), reference ≤2. With no startFrame the adapter may generate the first frame itself."
      properties:
        prompt: { type: string, minLength: 1, maxLength: 1000 }
        duration: { enum: [4, 6, 8] }
        resolution: { enum: ["720p", "1080p"], default: "720p" }
        aspectRatio: { enum: ["16:9", "9:16"], default: "16:9" }
        generateAudio: { type: boolean, default: false }
    VideoReframeParams:
      type: object
      additionalProperties: false
      required: [aspectRatio]
      properties:
        aspectRatio: { enum: ["9:16", "1:1", "4:5", "16:9"] }